if the file already exists, no file will be created. Otherwise, 3 files will be created. issues: #35

+++ level_1.py
@@ -1,20 +1,20 @@
+import os
 import pickle
 import random
 import sys
 import pygame
 from pygame.math import Vector2
 
-# create binary score file
-with open('scores_easy.bin', 'wb') as file:
-    pickle.dump([], file)
+if not os.path.exists('scores_easy.bin') and os.path.exists('scores_medium.bin') and os.path.exists('scores_hard.bin'):
+    # create binary score file
+    with open('scores_easy.bin', 'wb') as file:
+        pickle.dump([], file)
 
-with open('scores_medium.bin', 'wb') as file:
-    pickle.dump([], file)
+    with open('scores_medium.bin', 'wb') as file:
+        pickle.dump([], file)
 
-with open('scores_hard.bin', 'wb') as file:
-    pickle.dump([], file)
-
-score = 0
+    with open('scores_hard.bin', 'wb') as file:
+        pickle.dump([], file)
 
 
 class WALL:
